move service under org.oskari
